[1/2] cpu_input_boost: Rework display state awareness logic for OnePlus 8 devices
Currently, cpu_input_boost fails to engage when the device is
initially powered on until the display is manually cycled off/on.
Additionally, it occasionally fails to re-engage following
subsequent display power cycles. In order to ensure proper
functionality, we need to be more explicit about which display
states the driver is checking for.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Change-Id: I8ad61c4203b3653c67428bd44e433e5f436b217e